Utility - clearAll()

+++ utility/clear.js
@@ -1,5 +1,7 @@
 var Parse = require('parse/node');
 var config = require('./config');
+Parse.initialize(config.ParseApplicationId, config.ParseJavascriptKey, config.ParseMasterKey);
+Parse.Cloud.useMasterKey();
 
 process.argv.forEach(function(val, index, array) {
 	// https://nodejs.org/docs/latest/api/process.html#process_process_argv
@@ -12,23 +14,37 @@ process.argv.forEach(function(val, index, array) {
 // clears all Parse objects associated with given val.
 // val is the Parse Class
 function clearParse(val) {
-	Parse.initialize(config.ParseApplicationId, config.ParseJavascriptKey);
-	console.log("will clear " + val);
-	var ParseClass = Parse.Object.extend(val);
-	var query = new Parse.Query(ParseClass);
-	// find each obj and delete it
-	query.find().then(function(objs){
-		objs.forEach(function(obj){
-			obj.destroy({
-				success: function(obj){
 
-				},
-				error: function(obj, error){
-					console.log(obj);
-					console.log(error.message);
-				}
-			})
-		});
-		console.log("finished clearing all " + val);
+	console.log("will clear " + val);
+	if(val == "All") {
+		clearAll();
+	}
+	// delete only objects of the given class
+	else {
+		var ParseClass = Parse.Object.extend(val);
+		var query = new Parse.Query(ParseClass);
+		// find each obj and delete it
+		query.find().then(function(objs){
+			objs.forEach(function(obj){
+				obj.destroy({
+					success: function(obj){
+
+					},
+					error: function(obj, error){
+						console.log(obj);
+						console.log(error.message);
+					}
+				})
+			});
+			console.log("finished clearing all " + val);
+		})
+	}
+	
+function clearAll() {
+	var array = ["User", "_Session", "Debate", "Argument"];
+	array.forEach(function(a) {
+		clearParse(a);
 	})
+}
+
 }
